// src/types/event.ts
export interface Event {
  id: string;
  title: {
    en: string;
    zh: string;
  };
  date: {
    en: string;
    zh: string;
  };
  time: {
    en: string;
    zh: string;
  };
  location: {
    en: string;
    zh: string;
  };
  image: string;
  description: {
    en: string;
    zh: string;
  };
} 